+++ sdmx-dl-provider-base/src/test/java/_test/sdmxdl/util/XRepoRestClient.java
@@ -24,7 +24,9 @@ import sdmxdl.provider.Marker;
 import sdmxdl.provider.web.RestClient;
 
 import java.io.IOException;
+import java.util.EnumSet;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Stream;
 
 /**
@@ -77,8 +79,8 @@ public final class XRepoRestClient implements RestClient {
     }
 
     @Override
-    public boolean isDetailSupported() {
-        return true;
+    public Set<Feature> getSupportedFeatures() {
+        return EnumSet.allOf(Feature.class);
     }
 
     @Override